refactor: 그냥 int로 수정

+++ secretboard/src/main/java/com/siat/secretboard/board/dto/BoardRequestDTO.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class BoardRequestDTO {
-    private Long boardIdx;
+    private int boardIdx;
     private String boardName;
     private String regDate;
     private String updateDate;
